+++ lib/src/widgets/app_bars/minimalist_app_bar.dart
@@ -3,13 +3,16 @@ import 'package:lit_ui_kit/lit_ui_kit.dart';
 
 class MinimalistAppBar extends StatelessWidget implements CustomAppBar {
   final Color backButtonBackgroundColor;
+  final Color backButtonIconColor;
 
   /// The app bar height is depending on the abstract [CustomAppBar] height.
   static const height = CustomAppBar.height;
 
   const MinimalistAppBar({
     Key? key,
-    this.backButtonBackgroundColor = LitColors.mediumGrey,
+    this.backButtonBackgroundColor =
+        LitBackButtonDefaultStyling.backgroundColor,
+    this.backButtonIconColor = LitBackButtonDefaultStyling.iconColor,
   }) : super(key: key);
   @override
   Widget build(BuildContext context) {
@@ -30,10 +33,11 @@ class MinimalistAppBar extends StatelessWidget implements CustomAppBar {
                   child: Align(
                     alignment: Alignment.centerLeft,
                     child: SafeArea(
-                        child: LitBackButton(
-                      backgroundColor: backButtonBackgroundColor,
-                      iconColor: Colors.white,
-                    )),
+                      child: LitBackButton(
+                        backgroundColor: backButtonBackgroundColor,
+                        iconColor: backButtonIconColor,
+                      ),
+                    ),
                   ),
                 )
               : SizedBox()