+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Shorten the pubspec.yaml description's length to increase the pub score.
 - Add animation and adjust styling on LitLockIcon.
 - Refactor the ApplicationLicensesScreen
+- Add styling members to app bar widgets
 
 ## 1.0.4